fix review issues_2
